requirements.txt
sqlalchemy
google-generativeai
psycopg2-binary
scipy
soundfile
aiohttp
asyncio
pyworld
matplotlib
numpy
pandas
requests
websockets>=12.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
uvicorn
fastapi
